+++ packages/headless/src/features/facets/category-facet-set/category-facet-set-slice.test.ts
@@ -23,6 +23,7 @@ import {
   CategoryFacetSetState,
   getCategoryFacetSetInitialState,
 } from './category-facet-set-state';
+import {deselectAllFacets} from '../generic/facet-actions';
 
 describe('category facet slice', () => {
   const facetId = '1';
@@ -119,6 +120,17 @@ describe('category facet slice', () => {
     categoryFacetSetReducer(state, action);
 
     expect(FacetReducers.handleFacetDeselectAll).toHaveBeenCalledTimes(1);
+  });
+
+  it('dispatching #deselectAllFacets calls #handleFacetDeselectAll for every category facet', () => {
+    jest.spyOn(FacetReducers, 'handleFacetDeselectAll');
+
+    state['1'] = buildMockCategoryFacetRequest();
+    state['2'] = buildMockCategoryFacetRequest();
+    state['3'] = buildMockCategoryFacetRequest();
+    categoryFacetSetReducer(state, deselectAllFacets());
+
+    expect(FacetReducers.handleFacetDeselectAll).toHaveBeenCalledTimes(4);
   });
 
   it('dispatching #updateCategoryFacetNumberOfValues calls #handleFacetUpdateNumberOfValues if there are no nested children', () => {